requirements.txt
django
gunicorn
django-heroku
python-dotenv
django-allauth
termcolor
pyyaml
django-cors-headers
googlemaps
twilio
regex
stripe
